xen/arm: ffa: List VMs in part_info_get
List VMs in the system with FF-A support in part_info_get.
VM to VM functionalities in FFA support are only activated if
CONFIG_FFA_VM_TO_VM is activated in the configuration.
This patch is also introducing this new configuration parameter

WARNING: There is no filtering for now and all VMs are listed !!

Signed-off-by: Bertrand Marquis <bertrand.marquis@arm.com>